+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.15.2-dev
 
 - Add additional types at the API boundary (in `lib/parser.dart` and others).
-- Update to `package:lints` 2.0.
+- Adopted the `package:dart_flutter_team_lints` linting rules.
 - Fixed an issue with `querySelector` where it would fail in some cases with
   descendant or sibling combinators (#157).