+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Adopted the `package:dart_flutter_team_lints` linting rules.
 - Fixed an issue with `querySelector` where it would fail in some cases with
   descendant or sibling combinators (#157).
+- Add an API example in `example/`.
 
 ## 0.15.1